+++ R/utils.R
@@ -142,55 +142,60 @@ common_range <- function(x, stat_info) {
 #' Calculate Plot Labels for Numeric Missing
 #'
 #' Internal function that calculates information needed for modifying a numeric x axis
-#' to contain a value "NA".
+#' to contain a value "NA". This function can be improved. Note that the result is
+#' built consistently around ggplot2, not plotly.
 #'
 #' @noRd
 #' @keywords internal
 #'
-#' @param x Object of class "EffectData"
-#' @param plotly Plotly (`TRUE`) or not?
+#' @param mids Bin mid values.
+#' @param widths Bin widths.
+#' @param plotly Logical flag whether plotly's or ggplot's labeler is to be used.
 #' @returns
 #'   A list with the numeric representation of the NA value, the vector of numeric
 #'   breaks, and the vector of corresponding labels.
-numeric_scale_with_na <- function(x, plotly) {
-  # Note: the nrow(x) == 1 NA case was turned into character, and is thus
-  # not possible at this point
-  n <- nrow(x)
-  mids <- x$bin_mid
+numeric_scale_with_na <- function(mids, widths, plotly) {
+  n <- length(mids)
   stopifnot(
-    n >= 2L,
+    n == length(widths),
     is.na(mids[n]),
     is.numeric(mids)
   )
-  r <- range(mids, na.rm = TRUE)
-  breaks <- labeling::extended(r[1L], r[2L], m = min(4L, n - 1L))  # TODO find out plotly rule
-  # TODO: reduce to unique values of mids if length(mids) <= 3?
 
+  # Easy: the only bin is the NA bin
+  if (n == 1L) {
+    return(list(na.value = 0, breaks = 0, labels = "NA"))
+  }
+
+  # If no bars are to be plotted, we could even narrow the range further down
+  r <- mids[c(1L, n - 1L)] + c(-1, 1) * widths[c(1L, n - 1L)] / 2
+  breaks <- labeling::extended(r[1L], r[2L], m = max(2L, min(5L, n - 1L)))
+  m <- length(breaks)  # I hope that this has length >= 2
+
+  # Let's mimic the defaults
   if (plotly) {
-    # More or less the default scaler of plotly
     labeler <- scales::label_number(scale_cut = scales::cut_long_scale())
   } else {
     labeler <- scales::label_number_auto()
   }
   labels <- labeler(breaks)
 
-  m <- length(breaks)
-  min_gap <- sum(utils::tail(x$bin_width, 2)) * 0.6
+  # Minimal gap between r[2] and the midpoint of the NA bin
+  min_gap <- widths[n] * 0.7
 
+  # The last break is so far away that it can play the role of the NA break
   if (r[2L] + min_gap <= breaks[m]) {
-    na.value <- breaks[m]
     labels[m] <- "NA"
-  } else {
-    if (length(breaks) == 1L) {
-      D <- ceiling(min_gap)  # TODO: Better logic
-      nbreaks <- 1
-    } else {
-      D <- diff(breaks)[m - 1L]
-      nbreaks <- which(breaks[m] + D * (1:10) >= r[2L] + min_gap)[1L]
-    }
-    na.value <- breaks[m] + D * nbreaks
-    breaks[m + 1L] <- na.value
-    labels[m + 1L] <- "NA"
+    return(list(na.value = breaks[m], breaks = breaks, labels = labels))
   }
+
+  # How many additional breaks would we need to cover min_gap?
+  D <- diff(breaks)[1L]  # All diffs are equal
+  additional_breaks <- which(breaks[m] + (1:10) * D >= r[2L] + min_gap)[1L]
+
+  na.value <- breaks[m] + additional_breaks * D
+  breaks[m + 1L] <- na.value
+  labels[m + 1L] <- "NA"
+
   return(list(na.value = na.value, breaks = breaks, labels = labels))
 }